Rename requirement_checker to filters
And move filter configuration there.
